9 proyecto de autenticacion/requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
click==8.1.3
cloudinary==1.29.0
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
Flask-JWT==0.3.2
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.4.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
uwsgi